fix(preview): TWP-05 — 32-bit event indices, monotonic merge
The per-vertex TWP-frame, WCS-epoch and TLO/tool indices were u8 arrays
capped at 0xfe: a program with more than 254 events on a channel (a
multi-fixture program alternating G54..G57 per operation crosses it in
about 64 operations) silently posed everything after the cap with event
254's frame / fixture / tool. The three channels are now Uint32Array with
one sentinel (viewer/eventIndex.ts EVENT_NONE = 0xffffffff; TLO_NONE is
its alias; the WCS channel keeps 0 = epoch 0) through the decoder, the
scrub track, the split streams, the entry track, slices, both workers,
the part-frame transform, the collision track and the WCS rebase. The
resolver is also O(V + E): per-stream seq is ascending by construction,
so one forward pointer over the seq-sorted events replaces the nested
per-vertex scan; an out-of-order vertex falls back to a binary search on
its own. Client-only — no wire change, no schema bump.

Review: docs/reviews/twp-review-2026-09-14.md TWP-05.

Tests: previewDecode.test.ts "event indices are 32-bit" (257 events →
index 256 on every channel; exact event at every vertex across 1200
events; out-of-order vertex); existing strict/tie/fill tests kept with
the new constants; scrubTrack/collision/partFrame/wcsEpochs/trsrn
fixtures widened.

+++ lcnc-webui/src/previewWorker.ts
@@ -55,12 +55,12 @@ self.onmessage = async (e: MessageEvent<Req>) => {
     let rapidBreaks: Uint32Array | undefined;
     let feedMode: Uint8Array | undefined;
     let rapidMode: Uint8Array | undefined;
-    let feedFrame: Uint8Array | undefined;
-    let rapidFrame: Uint8Array | undefined;
-    let feedWcs: Uint8Array | undefined;
-    let rapidWcs: Uint8Array | undefined;
-    let feedTlo: Uint8Array | undefined;
-    let rapidTlo: Uint8Array | undefined;
+    let feedFrame: Uint32Array | undefined;
+    let rapidFrame: Uint32Array | undefined;
+    let feedWcs: Uint32Array | undefined;
+    let rapidWcs: Uint32Array | undefined;
+    let feedTlo: Uint32Array | undefined;
+    let rapidTlo: Uint32Array | undefined;
     // Outside-limits flags (2026-09-12): through the track split when there
     // is a track, else the wire arrays as they are (legacy strips).
     let feedOutside: Uint8Array | undefined = d.feed.outside;
